todo edit -ing : MenuServiceImplements.java

+++ src/main/java/com/kiosk/kioskback/service/implementation/CategoryServiceImplements.java
@@ -14,9 +14,9 @@ import com.kiosk.kioskback.dto.response.category.GetCategoryResponseDto;
 import com.kiosk.kioskback.dto.response.category.PatchCategoryResponseDto;
 import com.kiosk.kioskback.dto.response.category.PostCategoryResponseDto;
 import com.kiosk.kioskback.entity.CategoryEntity;
+import com.kiosk.kioskback.entity.StoreEntity;
 import com.kiosk.kioskback.entity.UserEntity;
 import com.kiosk.kioskback.repository.CategoryRepository;
-import com.kiosk.kioskback.repository.MenuRepository;
 import com.kiosk.kioskback.repository.StoreRepository;
 import com.kiosk.kioskback.repository.UserRepository;
 import com.kiosk.kioskback.service.CategoryService;
@@ -25,9 +25,8 @@ import com.kiosk.kioskback.service.CategoryService;
 public class CategoryServiceImplements implements CategoryService {
 
     @Autowired private CategoryRepository categoryRepository;
-    @Autowired private UserRepository userRepository;
     @Autowired private StoreRepository storeRepository;
-    @Autowired private MenuRepository menuRepository;
+    @Autowired private UserRepository userRepository;
 
     public ResponseDto<List<GetCategoryResponseDto>> getList(int storeId) {
         List<GetCategoryResponseDto> data = null;
@@ -45,21 +44,29 @@ public class CategoryServiceImplements implements CategoryService {
         return ResponseDto.setSuccess(ResponseMessage.SUCCESS, data);
     }
 
-    public ResponseDto<PostCategoryResponseDto> postCategory(String userId, PostCategoryDto dto) {
-        PostCategoryResponseDto data = null;
+    public ResponseDto<List<PostCategoryResponseDto>> postCategory(String userId, PostCategoryDto dto) {
+        List<PostCategoryResponseDto> data = null;
+
+        int storeId = dto.getStoreId();
 
         try {
 
             UserEntity userEntity = userRepository.findByUserId(userId);
             if (userEntity == null) return ResponseDto.setFailed(ResponseMessage.NOT_EXIST_USER_ID);
+            if(!userEntity.isAdmin()) return ResponseDto.setFailed(ResponseMessage.NOT_ADMIN);
+            
+            StoreEntity storeEntity = storeRepository.findByStoreId(storeId);
+            if(storeEntity == null) return ResponseDto.setFailed(ResponseMessage.NOT_EXIST_STORE_ID);
 
-            // todo : 로그인 유저가 관리자 인지 확인
-            // todo : 로그인 유저가 매장 주인이 맞는지 확인
+            boolean isEqualUserId = storeEntity.getUserId().equals(userId);
+            if(!isEqualUserId) return ResponseDto.setFailed(ResponseMessage.NOT_PERMISSION);
 
             CategoryEntity categoryEntity = new CategoryEntity(dto);
             categoryRepository.save(categoryEntity);
 
-            data = new PostCategoryResponseDto(categoryEntity);
+            // todo : 리스트를 다시 줘야하는게 맞는지 다시 생각해볼 필요가 있음.
+            List<CategoryEntity> categoryEntityList = categoryRepository.findByStoreIdOrderByCategoryPriorityDesc(storeId);
+            data = PostCategoryResponseDto.copyList(categoryEntityList);
 
         } catch (Exception exception) {
             exception.printStackTrace();
@@ -80,15 +87,20 @@ public class CategoryServiceImplements implements CategoryService {
             if (categoryEntity == null) return ResponseDto.setFailed(ResponseMessage.NOT_EXIST_CATEGORY_ID);
    
             UserEntity userEntity = userRepository.findByUserId(userId);
-            // todo : userId가 존재하는지 확인
+            if (userEntity == null) return ResponseDto.setFailed(ResponseMessage.NOT_EXIST_USER_ID);
             if (!userEntity.isAdmin()) return ResponseDto.setFailed(ResponseMessage.NOT_ADMIN);
 
-            // todo : 로그인 유저가 매장 주인이 맞는지 확인
+            int storeId = categoryEntity.getStoreId();
+            StoreEntity storeEntity = storeRepository.findByStoreId(storeId);
+            if(storeEntity == null) return ResponseDto.setFailed(ResponseMessage.NOT_EXIST_STORE_ID);
+
+            boolean isEqualUserId = storeEntity.getUserId().equals(userId);
+            if(!isEqualUserId) return ResponseDto.setFailed(ResponseMessage.NOT_PERMISSION);
 
             categoryEntity.patch(dto);
             categoryRepository.save(categoryEntity);
 
-            data = new PatchCategoryResponseDto(categoryId, userId, categoryId);
+            data = new PatchCategoryResponseDto(categoryEntity);
 
         } catch (Exception exception) {
             exception.printStackTrace();
@@ -107,14 +119,15 @@ public class CategoryServiceImplements implements CategoryService {
             if (categoryEntity == null) return ResponseDto.setFailed(ResponseMessage.NOT_EXIST_CATEGORY_ID);
 
             UserEntity userEntity = userRepository.findByUserId(userId);
-            // todo : userId가 존재하는지 확인
+            if (userEntity == null) return ResponseDto.setFailed(ResponseMessage.NOT_EXIST_USER_ID);
             if (!userEntity.isAdmin()) return ResponseDto.setFailed(ResponseMessage.NOT_ADMIN);
 
-            // todo : 로그인 유저가 매장 주인이 맞는지 확인
+            int storeId = categoryEntity.getStoreId();
+            StoreEntity storeEntity = storeRepository.findByStoreId(storeId);
+            if(storeEntity == null) return ResponseDto.setFailed(ResponseMessage.NOT_EXIST_STORE_ID);
 
-            // todo(논의점) : 카테고리 삭제시 메뉴를 모두 삭제할 것인지 남겨놓을 것인지
-            // todo(유호성 의견) : 메뉴는 남겨놓고 카테고리만 삭제하는 건 어떤지? menuEntity의 categoryId가 notNull이 아니면 가능
-            menuRepository.deleteByCategoryId(categoryId);
+            boolean isEqualUserId = storeEntity.getUserId().equals(userId);
+            if(!isEqualUserId) return ResponseDto.setFailed(ResponseMessage.NOT_PERMISSION);
 
             categoryRepository.delete(categoryEntity);
